+++ UseBy/Controllers/Profile/ProfileViewController.swift
@@ -9,6 +9,7 @@ import Foundation
 import UIKit
 import SnapKit
 import MessageUI
+import GoogleSignIn
 
 class ProfileViewController: UIViewController {
     struct UIConstants {
@@ -69,6 +70,7 @@ class ProfileViewController: UIViewController {
             make.height.equalTo(UIConstants.heightTableView)
         }
     }
+
     func configureLogOutButton() {
         view.addSubview(logOutButton)
         logOutButton.snp.makeConstraints {(make) -> Void in
@@ -77,6 +79,14 @@ class ProfileViewController: UIViewController {
             make.centerX.equalTo(view)
             make.top.equalTo(profileTableView.snp.bottom).offset(UIConstants.spaceBetweenTableAndLogout)
         }
+        logOutButton.addTarget(self, action: #selector(didTapLogoutButton), for: .touchUpInside)
+    }
+
+    @objc
+    func didTapLogoutButton() {
+        GIDSignIn.sharedInstance()?.signOut()
+        let mainViewController = MainAuthViewController()
+        self.view.window?.rootViewController = Router(rootViewController: mainViewController)
     }
 }
